// deploymentanalyzer.client/src/app/models/mastercode.ts
export interface MasterCode {
  id: number;
  name: string | null;
  category: string | null;
  added_by: string | null;
  added_date: string | null;
  region: string | null;
  currency_symbol: string | null;
  country: string | null;
  company: string | null;
  added_by_name: string | null;
  added_dt: Date | null;
  website: string | null;
  headquarters: string | null;
  number_of_employees: number | null;
  type_of_entity: string | null;
  sales_previous_year: string | null;
  salesforce_name: string | null;
  generic_name: string | null;
  us_name: string | null;
  product_name: string | null;
  year: number | null;
  quarter: number | null;
  type_of_salesforce: string | null;
}